requirements.txt

base58==2.1.1
based58==0.1.1
solana==0.23.3
django-environ==0.9.0
